Add files via upload
Mastering Ethical Hacking: A Daily Guide to Live Methodologies is a 
comprehensive handbook that takes you on a day-to-day journey through the 
exciting and ever-evolving world of ethical hacking. Whether you're a novice looking 
to dive into the cybersecurity field or an experienced professional seeking to 
enhance your skills, this book is your go-to resource.